src/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
blinker==1.9.0
certifi==2025.1.31
click==8.1.8
distro==1.9.0
dotenv==0.9.9
exceptiongroup==1.2.2
faiss-cpu==1.10.0
Flask==3.1.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
MarkupSafe==3.0.2
numpy==2.2.3
openai==1.65.2
packaging==24.2
pluggy==1.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pypdf==5.3.1
pytest==8.3.5
python-dotenv==1.0.1
sniffio==1.3.1
tomli==2.2.1
tqdm==4.67.1
typing_extensions==4.12.2
Werkzeug==3.1.3
